pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sns-automation"
version = "0.1.0"
description = "SNSアカウント構築・運用マニュアル自動化システム"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "SNS Automation Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.39.0",
    "google-auth>=2.27.0",
    "google-api-python-client>=2.114.0",
    "elevenlabs>=1.2.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pillow>=10.2.0",
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[project.scripts]
sns-automation = "sns_automation.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
